# blackjack-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blackjack API",
        "description": "This is a simple Blackjack API. For more information, please visit the developer's portfolio at [Ignacio Albiol's Portfolio](https://ignacioalbiol.tech/).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Blackjack API Server"
        }
    ],
    "paths": {
...
        "/players": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show all players & their average win percentages",
                "description": "This endpoint returns all players and their average win percentages. Only authenticated users with the appropriate permissions can access this endpoint.",
                "operationId": "getAllPlayers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_wins_average": {
                                            "type": "number",
                                            "example": "0.52"
                                        },
                                        "total_losses_average": {
                                            "type": "number",
                                            "example": "0.30"
                                        },
                                        "total_ties_average": {
                                            "type": "number",
                                            "example": "0.18"
                                        },
                                        "user_details": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register a new player",
                "description": "This is the endpoint to create and register a new player.",
                "operationId": "registerPlayer",
                "requestBody": {
                    "description": "Player registration details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nickname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Player created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Player created successfully"
                                        },
                                        "uuid": {
                                            "type": "string",
                                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "uuid": {
                        "description": "The unique identifier of the user",
                        "type": "string"
                    },
                    "nickname": {
                        "description": "The nickname of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email of the user",
                        "type": "string"
                    },
                    "wins": {
                        "description": "The number of wins of the user",
                        "type": "integer"
                    },
                    "losses": {
                        "description": "The number of losses of the user",
                        "type": "integer"
                    },
                    "ties": {
                        "description": "The number of ties of the user",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication",
                "name": "Bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}